+++ package.json
@@ -6,18 +6,22 @@
   "author": "Alistair Smith <alistairsmith01@gmail.com>",
   "license": "MIT",
   "devDependencies": {
-    "@now/node": "^1.8.3",
     "@types/moment": "^2.13.0",
-    "@types/node-fetch": "^2.5.7",
-    "prettier": "^2.1.2",
-    "vercel": "^20.1.2"
+    "@types/node-fetch": "^3.0.3",
+    "prettier": "^2.4.1",
+    "vercel": "^23.1.2"
   },
   "dependencies": {
-    "discord.js": "^12.4.1",
+    "@linear/sdk": "^1.20.0",
+    "@vercel/node": "^1.12.1",
+    "discord.js": "^13.1.0",
     "moment": "^2.29.1",
-    "node-fetch": "^2.6.1"
+    "next": "^11.1.2",
+    "nextkit": "^1.1.0",
+    "node-fetch": "^3.0.0",
+    "zod": "^3.8.2"
   },
   "scripts": {
-	"d": "vercel dev"
+    "d": "vercel dev"
   }
 }
